no schema means better TS ;)
